Fix: Chris van Engelen <engelen@lucent.com> states:
SourceView::full_path attempts to remove parts of the filename that
look like "/dir1/../". However, in doing so, it also removes the
string "/../../" from the filename, causing it to make the filename
incorrect!!!

Furthermore, strings like "/dir1/../" or "/./" should be removed
repeatedly.

I have tested this code and the bug is fixed.