Matthew Persico <Matthew.Persico@lazard.com> wrote:
DDD has problems evaluating Perl hashes and hash references. One of
the problems is that DDD does not quote the key values. If the key
leads with a digit or has non-alphanumeric characters in it, the
inferior perl debugger reports an error and the graph displays do not
evaluate correctly.

Now fixed.