New: forwarding of keyboard events from source window to console is optional;      controlled by the `consoleHasFocus' resource.
Colin <cbyh@yahoo.com> writes:

We are running DDD via its tty interface so that it can be driven by
commands from a parent process. When DDD is started up in this way
(via the --tty option) it by default hides the GDB command pane.

However even though the pane is hidden, you can still type commands
and they are executed by DDD (even though you can't see what you are
typing!)

We think that this is wrong. Whenever the GDB pane is closed then all
keyboard-based command input should be disabled as well.

This causes us a particular problem because we are sending commands to
DDD via its tty interface, but if the user accidentally hits the
keyboard then the keypresses get interleaved with the tty commands and
we end up with a complete failure due to unrecognised commands!