Patch by Peter Ernst (5 Jun 2002)
DDD 3.3.1 with Perl: when using the graphical display on reference
variables, de-referencing works nicely through e.g. a double-mouse
click (by means of code in ddd/deref.C). However, if the references is
blessed (i.e. it is a object reference), de-referencing does NOT work.

Example: graph display $SeqInfoObj

p $SeqInfoObj
W3H::Parser::Seqinfo=HASH(0x4d3738)

i.e. $SeqInfoObj is a blessed reference

double-click on the graph does NOT dereference, because
the code in ddd/deref.C tries to get the "type" of the
reference (in this case HASH) by parsing the output of
" p $SeqInfoObj ", and considers the first identifier as
"type". In this case the type is "W3H" which is wrong.

Interestingly, the regular expression to parse the output
of " p $SeqInfoObj " (rxperlref line 359 in ddd/rxscan.L)
already knows about the "more complicated" syntax. I.e. it
is not necessary to fix something here.

Fix description:

Correct is is to take the first identifier after a "=".
This works for blessed and un-blessed references (un-blessed
references don't have a "=" character in the output of " p $simple_ref ").
I.e. the bug is fixed by changing line 72 in ddd/deref.C
from:
ref = val.through(rxidentifier);
to:
ref = val.at(rxidentifier, val.index("=") + 1);

On a blessed reference, val.index("=") + 1 points to the first
character after the "=" sign,
on a un-blessed reference, val.index("=") + 1 points to character 0.

The fix is working as you can see from the history file below:
$SeqInfoObj is a blessed hash-reference

$SeqInfoObj->{'entries'} is an un-blessed array-reference

$SeqInfoObj->{'entries'}->[0] aka
${$SeqInfoObj->{'entries'}}[0]
is an un-blessed hash-reference